Wouldn't it be great if I could relate all the info in the world in a way that is understandable?
If I had some stock like RHM, I wish I could look into that system and say how it relates to all the other stocks that have ever been in that situation in the past. But not just price info, but all the relevant factors in the real world as well. If I could create such a system, trading would fall like a domino. It's all a matter of having sufficient intelligence rather than operating blindly. For now, it'd be great if for swing trading I could gather the necessary data and at least determine what the right entries for such trades are. For mean reversion longs and shorts. What puts me off mean reversion trading is that I could never quite figure that out.

I am really not into systems that much. I'd rather have a big database over which to do a similarity search. Then once I understand the possible scenarios based on historical data, then the trading plan for that specific situation can be formulated. That is how I want to trade. Not via indicators. But via memory.