Putting so much exercise into execution doesn't sit right with me, especially given that my RSI hasn't healed. It's not a skill that I need. The more I can substitute with programming the better off I'll be. Let the work of the machines be the work of my own hands.
When you read The Playbook and now Lance's course, you really get the sense just how much effort trading really requires.

I don't want to be a soldier or play an FPS like they do. I want to be the general sitting behind the lines in a cushy office and giving orders. I'll still have to work towards it, but it isn't the same path.