And if I fail, so be it. I'll accept my defeat. Making a lot of money as a trader is besides the point.
I want more than anything else for the world to accept me.

I don't want the world to accept the prop trader me who puts in 60h/week trying to do it in a discretionary way.

I want it to accept my pursuit of the Singularity. I want it to recognize this effort.

I want...my own Way.