I am using a volume screen now instead of the 6 month momentum. Maybe that will work better. 95% of the stocks on this volume screen are trash. Maybe I should just filter out all the stocks below 5$?
I am a bit turned off from the daytrading system because I'd have to filter the stocks it trades manually. If it tried trading the shit stocks on this list, it would go broke quickly.

For each stock, I need to actual think about the pattern and the news catalyst associated with it.

Ironically, while I was modeling the probabilistic program for the daytrading system in my mind, that is when I realized the importance of volume. A change in volume can indicate a regime change in a stock, so it's good to have on your side.