It really matters a lot that I get good entries near the open. If I make 300 trades per year and improve my entries on average by 1% just how much would that add to my bottom line? I don't think trading breakouts at the close would be a good idea for that reason. Which is why I didn't do it.
I was just too damn lazy to win in the 2007 - 2014 era.

This is the main issue I need to fix now. I need to stop looking for shortcuts and take advantage of the opportunity as it exists. I need to work for my money. I need to make trading my job.