I am lacking sleep, overthinkimng it and working myself into a frenzy. I shouldn't trade breakouts as if I were a daytrader. Q is right. just put the stop at the low of the day and size the position down if it's overextended or don't trade at all. Rather than asking when to buy and sell I should be asking how much.
Trading breakouts isn't rocket science. The breakout above 4 could have been a good trade if sized appropriately and given room to move. Even buying at 4.3 is fine. The stock moves insanely fast. 5-10% in it is not that much.

I need to stop thinking that I need to allocate 10% of my account on every trade. I am being a moron.

A 10% in one stock might be nothing, but a 10% in another might be insanely risky.

I know this, so why am I approaching every problem by thinking of putting in 10%?