No more trades. I am dropping trading from consideration. I either do it properly or I not do it at all.
Any bots I create would only end up simple. The ORB bot for example would not be much different from a separate order type. It's not a big deal at all.

The preparation necessary to swing trade, and the constant chart reviews and early morning trading might not take as much time as daytrading does, especially if I create bots to help me do it, but it would still be a full time job. And when I look deep inside of me, the pain in my gut is telling me that I am not ready to commit.

I have no psychological risk capital left.

Instead of aiming to become a 3/6 trader, what I should do is fulfil my destiny as a 6/6 programmer and finish that poker bot. Go back to programming in Spiral and just do it. Master GPUs, and Cutlass and bring the project to its conclussion.

That could be worth millions per year in income at a very low risk.