Giving up on being a 'trader' is the right choice. Who cares about the big money? I should just care about the easy money. And focus on pursuing the Singularity.
I already didn't want to become a trader from the start, so what is the point of stressing myself out over it? I was only ever interested in the money it could make me.

The answer is to find the easy money trades, only take those and be done with it.