That was a random application, but I got the job at Tao Ceti in a similar manner, through a Reddit post when I applied to it on a whim. The spraying and praying approach on LinkedIn and elsewhere really got me nowhere for a long time, and when I applied to Tao Ceti I wasn't even looking for one.
It's amazing that my resume does have some overlap with their requirements right now. Who knows, maybe I will get a bite? Most likely not. But doing moves like these to improve my chances in life feel good. The biggest issue is that they are in the US and I am in Europe and they are looking for somebody in the same timezone, so most likely this won't fan out.

I have that job with Thomas lined up anyway.