Yeah, you know what. Forget the SemiAnalysis job, I won't apply to anything else anymore. I have 100k. If I get on that job with Thomas, that will be my last pro software work. Either way, after it's done, I'll close the company and either poker bot or trade bot full time. Now that I've been given the setups, I am not going be a losing trader anymore.
I have the 2 simple ones on stocks in play. I also have the trend day setup in the futures market that I could attempt to quantify. That should be enough to get me started.

I have huge amount of leeway to give poker botting a try and then return to trading afterwards if that fails to pan out, or if I get tired of poker botting.