...Damn, I want to trade. I keep thinking about VWAP-ing those Stocks In Play. Forget that. I need to think about crushing fish in poker.
**God, I want to crush fish in poker!**

That is what I should be thinking here if I want to bring this project to a fruition.

Once I desire to play poker, I'll get a further boost of motivation. I should learn from Thomas. He logged into the poker room for the first time in ages and crushed some noobs with +EV.

That is the right demeanor to have.

Just think of that smooth line going into millions. You can do it.

Think about how much money you can make playing cards and go for it. If people can win in the markets by using bots, why couldn't I do the same in gambling?