I'll abandon the pursuit of Singularity in favor of the pursuit of machine rationality.
Just like I once abandoned the dream of omnipotence in favor of the dream of entering the self improvement loop, so will I once again change.

The self improvement loop is the Singularity.

It's too hard for me to pursue.

So the only answer is to cultivate my understanding of probability theory through the use of machines.

And there is no better game for that than poker.