![NBIS](images/image-677.png)
I had a feeling if it could go up above 92.5 which was the offering price, that it could explode. It seems I was right.

I don't really care about this right now. It's just curiosity.

If I take what RNG does personally, then I'll be really fucked.

That the market is a RNG or rather, a probability distribution doesn't mean it doesn't have patterns. It very well does. It's just that I should forget the ticker names and trade only those patterns.