If the solution to finding an edge is creating a model of the market, then wouldn't the solution to the problem of transcendence be to create a model of intelligence?
Probabilistic programming is not the same thing as intelligence. It's more like a framework for building it.

The issue of the inference algorithm for such a model aside, what would the model program be like? How would one even attempt to write it?

I've never really said to myself that probabilistic programming is the axiom and then tried to reason from that point. There are plenty of good reasons for it, but if I suspended my disbelief, how would one attempt to predict the world? If personality is genetically determined, how would one program the personality of a being? Its instincts?