Living in abundance means thinking:
I did it once and I can do it again.
Just because I haven't done it doesn't mean that I can't.
There is enough time to do what I truly want.

I need to believe that I have the time, the skill and the money. That the failures aren't defeats, but temporary setbacks.

I just need to put in the effort.