All this regret is caused by one thing - not finishing the poker bot. I have to accept that punishment and complete the damn thing.
And it cannot be helped that I have to go from the top. It's like a roguelike. I died and went to the main menu.

I'll get started on this tomorrow. I don't care about getting rid of the hangups. I'll just accept these feelings and swim upwards. That is my way.