I am imagining exactly what I need.
Something like the TradingView screener.

During the first 10 mins, I shouldn't even trade according to charts. If I'm properly prepared then reading the charts would just waste my time.

I just a need a watchlist with all the tickers there, and all the important data such as:

* Premarket relative volume
* Premarket gap %
* Relative volume
* Change % relative to previous day high
* Change %
* Price
* Daily high
* Daily low
* Avg daily Volume
* 14 day ATR
* 14 day ADR

I'll also need a way to quickly place trades with a single click. I don't want to be wasting my time entering them manually.

Will probably want some buttons on the watchlist for that. 1-click + Enter to buy a certain % of my account calculated automatically.

I need two separate watchlists. One would be 'To Buy' and the other 'To Watch'.

I'd only buy from the first one.

Yeah, this is the way to do it.

I need to setup absolutely everything so that when the opportunities are there, I'd be the first one to take advantage of them. Speed comes from preparation.

I absolutely don't want to run my screeners in the first 10 mins of the day and trade based off of that. That'd be ridiculous. I need to plan what my buy points are ahead of time and just act once they are triggered.

So with that as a goal, I need a platform that could fulfil those needs.

Let's not hurry with this. I'll take my time to set things up properly.