What's the play here?
When it comes to market panics, they happen in daily cascades, so to avoid them you have to sell into small dips and buy in 2 days later. The problem is that this kind of activity is negative EV.

Statistically, buying into 2.7 declines in the S&P like the one on Friday is +EV.

But the problem is that 20-30% of the time you get killed as the market keeps falling. It's also hard to figure out where to put stops if you're just catching falling knives.

So yeah, unless you're a huge trader, there are easier trades out there. I'll focus on the gap and go trades when the market recovers and see where that gets me.

With the BTC liquidation, that could cause the panic to spread, so I am just fine waiting it out.

In a bull market there should be one good gap and go setup a day.