Time for a lesson. Let me show a chart.
![CROX](images/image-1215.png)

This was a chart that originally gave me a lot of trouble, but is now completely transpared to me.

![CROX](images/image-1216.png)

In particular note this part here. Note how in the easy part of the upmove the stock was continually making new higher lows on the weekly for 20 bars straight.

And then once that pattern was broken, the stock still continued going up, but the 2 week system starts to be continually violated.

The stock would break the 2 week low, make new highs, and then break the 2 week lows again. Even though the uptrend was still intact by some definition, the momentum was dying. And then the stock finally cracked.

I finally understand it. I want to be in a stock when it smoothly making higher lows on the weekly charts, and I can just skip the hard part of the upmove.

I should be trading weekly trends, not the monthlies.

That is my edge.

I finally get it!

Lance Breitstein made a huge impact on me. I think something just clicked when I saw him use a 2 minute bar trend following system in the right contexts.

I think I'll be able to take it easy in the stock market from here on out.