I used to read charts like these as a beginning trader and though that this is impossible, nobody can trade it.
This is because I'd look at what happens when the stocks close at their high. I got it hammered into me in those early books that I read that stocks should be bought when they make new highs. This did a lot of damange. Darvas, O'Neil, Livermore all mislead me.

What I should have done is taken a step back and realized that the higher lows on the weekly are really what matters.

The trouble is that back then I had no idea of the market context or how to judge it.

There are many ways to define an uptrend, but in my opinion the best one is to take a timeframe like the weekly, and then look for a series of higher lows. I am talking about sequential bars, not arbitrary chart patterns.

That's also the simplest way to do it.

There is no confusion in how long to hold or when to get out. Even if I get transported in the early 20th century without computers I could still easily trade this system.