The longs got destroyed.
Imagine what would've happened had I been using my old 8 - 12 week system on this. This is why I don't want to use it anymore.

Yes it would have rode up BTC all the way, but it would also ride down the declines like these. It would catch all the crashes.

But even the 2 week system would have protected me on INT. Switching to the 1 week system on the final narrow week would have been ideal.

Some stocks with weak momentum tend to poke the lows, but even the 1 week system would work on stocks with strong momentum. Ultimately, I need to be flexible and switch between the 2 week, 1 week and 1 day as the stock accelerates to the upside.

A good rule of the thumb would be 8. 8 up weeks and 8 up days to switch.