Imagine I shorted here with a stop at the previous day's highs. Not only would I get stopped out, but if I shorted the exact stop afterwards it wuold have taken ages for the trade to turn profitable.
Whatever money there is in volatility trading doesn't come from timing the spikes.

It was just an idea. A few days ago I saw the 2020 rise and I thought that maybe shorting the VXX wouldn't have been a bad idea, but it's just be a waste of time even if I could short at the exact highs ahead of time.

I see that now.