It's not beyond reason to expect a correction at this juncture in the market, but a bear market is a far bigger thing to anticipate. They don't come around just because you want them.
I'll change my mind if more bank busts happen. Those are a sure sign of an upcoming bear market.

The froth in the market needs some time to cool down, but the most likely scenario is that the strong momentum market will resume after that. I'll have my chance to make a lot of money.