I think the ultimate decising factor is that hard stops have no abiguity to them.
Imagine I was trading a stock 10 minutes to the Friday close, the stop is 10.5 and the price is 10.49. 5 minutes later it 10.51. 3 mins later it 10.48. 1 min later it's 10.5 exactly.

What the fuck do I do in that case if I want to sell on the close of the bar?

With a hard stop, there is zero ambuguity. If the price ever crosses the stop price, the market order gets send out, no doubt whatsoever.