If I am going to beat the market I need to stop looking at prices moving and start reading the news. Anticipate the news flow. Anticipate how the others will react to it. That is the way to go to the next level.
For example, I am currently brearish on AI, but suppose a new algorithm gets discovered which works much better than existing ones? That could immediatelly flip the whole scenario to bullish.

AI is in a bubble and such a algorithm is very unlikely to get discovered in a timely manner, but the chance of it happening isn't 0 and is 100% in the long term.

This is one of the ways in which the AI bubble is different from all the other bubbles. None of the bubbles in history had a possibility of magically becoming much more valuable by simple knowledge improvements. A single paper being published could add trillions of dollars of value to the market.

This is a difference from any other time in history.

https://youtu.be/LeDXXOHsUYI
Most reliable pattern in stocks

This is probably Pradeep's most important video.