Thanks to my fans for sponsoring Simulacrum: Heaven's Key. Sorry for not writing in a while.
I'll keep this here for the time being, in case I need it. Nwm. I thought that I might need to transfer it out via Paypal. Hopefully, I'll be able to since it's Patreon at some point.

https://youtu.be/pNFIG0wWpho
Implementing The Enumerative CFR Algorithm For Leduc Poker In Spiral's Host Cuda Backend

Incidentally, here is the link to the latest screencast. I keep forgetting to post these in the journal.