Oh, my god, when will this webinar end? 25m long sales pitch.
For the record, I am convinced of the value of the scalp radar. If I weren't working on the pokerbot, I'd jump on this.

Bookmap, scalping.

This is something I couldn't even dream of 2007 - 2014 because I was so undercapitalized.

Right now I am hung up on waiting for that panic so I could go long, but if I have to stay out of the market for months just to find an opportunity, just how much could the weekly momentum strategy be worth?

Having good RR trades come up several times a day is much better than it happening several times per year.