Today the weather was awful. And because of my back pain I had trouble sleeping. Let me just finish the benchmark run and go through the Eigen docs and I will call it a day.
The stock market torments me so. Long term trading really is the worst. If I had a daytrading system, I wouldn't be worrying about taking time off. Yeah, it's ridiculous. It's really easy to see big winners that broke out in hindsight, but in practice there are a tons of breakouts which give only mediocre performance.

I'd rather work for my money than have to rely on the market conditions or any of that crap.