This is pretty easy, but these kinds of tasks are ideal for familiarizing myself with the platform.
For the actual bot, I'd probably want to host it on a public cloud that's closest to the exhange and run it there. But in that case it'd also have to run through the broker. I bet IBKR will have something close to the exhange suitable for high frequency trading.