Yeah, I am not really inspired by trading. In 2025 I came up with all these plans, but it feels like I was just getting a shitload of volatility. So I am not interested in weekly momentum anymore.
What pissed me off the most is the complete lack of skill and control over my equity curve. I just want to be a lot better at trading. In the next week, and by the end of the month, let me get out of the tutorial stage and into real trading, the kind that I couldn't have done back in 2007 - 2014.

Thomas inadvertedly is bailing me out in this trading venture. It would have been a lot more annoying doing all this without the agentic coding tool he introduced me to.