https://youtu.be/ixo_DoBGYko Bookmap Education Course Part 2 | Volume Profile and Microstructure
Finally I resume from a few months ago. Let me watch it for 30m and then it's breakfast time.

I should do my best. There is a lot of room for me to grow in this field. The ceiling is a lot higher than in poker. And unlike in poker I won't be bashing my head against an RL problem. I've learned not to do that anymore.