Yeah, forget this. I am not going to get this approach.
Abnormal volume.
Abnormal price action.
Abnormal price cascades.
Abnormal market breadth.
Abnormal sentiment.
Abnormal earnings.

All of this can create an edge. What I am doing with the ORB system is combining first with a little bit of the second.

Lance like to trade price cascades, and I am going to be learning that as well.

Forget reading the S&P. Looking at those Bookmap charts, it still as hard as ever.

I'll focus on mastering the first 3 and then I'll try to figure out the trend days in the market using 4. If I can do that I'll have my business.

If I can get a foothold on the open, I'll bring in probabilistic programming and try to get more nuanced in order to juice my returns. Let me read the One Good Trade book.

I know what I want to do as a trader. I want to bet it all on the power of the abnormal. If I want to take advantage of that, first I need to take a step back and take stock of the market. I will do that by building out the infra needed to act as my eyes and ears.