I really am not a good fit for a trader, so it's no wonder I couldn't win back then. But who cares about fit. I just need to do the research, make the plan and then execute it.
What was keeping me from profits was my comprehension. I put myself in a bad position to get it, and the rest was history. Strong determination won't help me if it's not applied in the right place.

Yeah, I am inferior, but this time I have the power of the machines on my side. It will be different.