> This is what we're waiting for. We're waiting for the level to develop. > This is becoming a little bit of a resistance level. The more he sells the bigger the resistance level.
Yeah one of the techniques demonstrated could be emulated with footprint charts, but not all of them. Footprint charts would be great for finding volume at certain price, but not good for figuring out when a buyer drops for example. In fact, I am not sure I could do the later even with a heatmap.

But well, it's not like Jesse Livermore could see the order book in the bucket shop. He was reading times and sales there.

Just by incorporating the insight of how much volume it usually takes to move a stock into my trading could develop into a significant edge.