I'll take some time to think how to design my strategy around this. Ironically, the best way forward is like how it has been so far. Implement the ORB using 5 min breakouts.
But afterwards I will have to see whether there is a way to improve it using tape reading concepts.

It's an edge, but it definitely isn't trivial to utilize.

Nevermind the drills. I am not willing to spent all my time in front of the charts right now. I want to do programming and quant research.

First, I'll find a way to pay my bills normally and then build upon that afterwards.